fixing JENKINS-47631: Plugin cannot access SV project on Jenkins slave - moving new and internal SV classes to sv specific package: com.microfocus.application.automation.tools.sv
